requirements.txt
# Enterprise Codebase Review Assistant
streamlit>=1.32.0
chromadb>=0.4.24
google-generativeai>=0.8.0
PyGithub>=2.2.0
# tree-sitter 0.22+ breaks tree-sitter-languages 1.10 wheels (Language API change)
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.6.0
